% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Traing the prediction model, calculating covariance matrix and updatinf the database.}
\usage{
build_model(train_data, DB, cov_data)
}
\arguments{
\item{train_data}{A matrix contains information about the predictors for the standard. The data frame should have the following columns: RT, logD, pos, Phosphates, HBD_MW, neg, Rotarable_Bonds.}

\item{DB}{The user's metabolite database.}

\item{cov_data}{A data frame containing the covariance data.
The data frame should have the following columns: Name, Predicted.RT, measured.RT, measured.mz, theoretical.mz, adduct.}
}
\value{
The covariance matrix, the standard deviation of the models residuals and the updated database.
}
\description{
This function creates the covariance matrix based on the given data.
}
\seealso{
trainModel create_COV_matrix build_model
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
