% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_COV_matrix.R
\name{create_COV_matrix}
\alias{create_COV_matrix}
\title{Creating the Covariance Matrix}
\usage{
create_COV_matrix(sd_LogRT, cov_data)
}
\arguments{
\item{sd_LogRT}{The standard deviation of the residuals of the model.}

\item{cov_data}{A data frame containing the covariance data.
The data frame should have the following columns: Name, Predicted.RT, measured.RT, measured.mz, theoretical.mz, adduct.}
}
\value{
The covariance matrix.
}
\description{
This function creates the covariance matrix based on the given data.
}
\seealso{
trainModel create_COV_matrix build_model
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
