% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\name{trainModel}
\alias{trainModel}
\title{Training Linear Regression Model}
\usage{
trainModel(data)
}
\arguments{
\item{data}{A matrix contains information about the predictors for the standard. The data frame should have the following columns: RT, logD, pos, Phosphates, HBD_MW, neg, Rotarable_Bonds.}
}
\value{
A trained Linear Regression Model.
}
\description{
This function trains a Linear Regression based on the standards that the user provides.
}
\seealso{
trainModel create_COV_matrix build_model
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
