% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDB.R
\name{updateDB}
\alias{updateDB}
\title{Update the database with the new predicted LogRT}
\usage{
updateDB(model, DB)
}
\arguments{
\item{model}{A trained prediction model}

\item{data}{A matrix contains information about the predictors for the standard. The data frame should have the following columns: RT, logD, pos, Phosphates, HBD_MW, neg, Rotarable_Bonds.}
}
\value{
The updated database.
}
\description{
This function updates a user defined database with the new prediction for LogRT.
}
\seealso{
trainModel create_COV_matrix build_model
}
\author{
Francesco Del Carratore \email{francescodc87@gmail.com}
}
